+++ src/pdfdeal/Doc2X/Exception.py
@@ -1,0 +1,5 @@
+class RateLimit(Exception):
+    """
+    达到每分钟最大请求次数
+    """
+    pass
